requirements.txt
numpy==2.4.0
opencv-python==4.13.0.90
nnfs==0.5.1
matplotlib==3.10.8
pillow==12.1.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
packaging==26.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
